pyproject.toml
[project]
name = "gh-events"
version = "0.1.0"
description = ""
authors = [
    {name = "Raviraj D"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "fastapi (>=0.116.0,<0.117.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "duckdb (>=1.3.2,<2.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "streamlit-autorefresh (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
